RAML Walker test using _examples/versions router
